Separating state from UploadAttachmentsWorker
