+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/channel/controller/WhenHandleEvent.kt
@@ -22,6 +22,7 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.offline.SynchronizedCoroutineTest
 import io.getstream.chat.android.offline.message.attachments.internal.AttachmentUrlValidator
 import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelLogic
+import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelStateLogic
 import io.getstream.chat.android.offline.plugin.state.channel.internal.ChannelMutableState
 import io.getstream.chat.android.offline.plugin.state.global.internal.GlobalMutableState
 import io.getstream.chat.android.offline.randomChannel
@@ -99,7 +100,7 @@ internal class WhenHandleEvent : SynchronizedCoroutineTest {
             globalMutableState,
             repos,
             false,
-            attachmentUrlValidator
+            ChannelStateLogic(channelMutableState, globalMutableState)
         )
     }
