+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/channel/controller/WhenUploadAttachmentsTests.kt
@@ -322,7 +322,7 @@ internal class WhenUploadAttachmentsTests {
 
         fun get(): UploadAttachmentsWorker {
             return UploadAttachmentsWorker(
-                logic = logicRegistry,
+                channelsManager = logicRegistry,
                 messageRepository = messageRepository,
                 chatClient = chatClient,
                 attachmentUploader = uploader
